trig: Use floating point versions of trig functions
Replaces calls to double-precision floating point trig functions with
calls to single-precision floating point trig functions as follows:
    cos() -> cosf()
    sin() -> sinf()
    tan() -> tanf()
    acos() -> acosf()
    asin() -> asinf()
    atan() -> atanf()

This allows the compiler to avoid generating unnecessary instructions
that convert a float to a double and then a double back to a float.

+++ code/asteroid/asteroid.cpp
@@ -1566,8 +1566,8 @@ void asteroid_maybe_break_up(object *pasteroid_obj)
 						float r = sqrt(1.0f - y*y);
 						float phi = i * inc;
 
-						dir_vec.xyz.x = cos(phi)*r;
-						dir_vec.xyz.y = sin(phi)*r;
+						dir_vec.xyz.x = cosf(phi)*r;
+						dir_vec.xyz.y = sinf(phi)*r;
 						dir_vec.xyz.z = y;
 
 						// Randomize the direction a bit
